requirements.txt
pycryptodome==3.9.7
pelican==4.2.0

